requirements.txt
asn1crypto==0.24.0
Authlib==0.7
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
Flask==1.0.2
future==0.16.0
gspread==3.0.0
gunicorn==19.8.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.5.0
MarkupSafe==1.0
pycparser==2.18
requests==2.18.4
six==1.11.0
tinytag==0.18.0
urllib3==1.22
Werkzeug==0.14.1
zhconv==1.4.0
